Revert "Replaced iron with magma"
Magma cannot be used at the moment because there is no way to modify the
Python path such that it points to the site packages of a virtual
environment. Also it seems like the Python that Magma uses is configured
differently than the on used in a shell such that it doesn't have access
to the Python packages installed in the dev shell.

This reverts commit 85ee62f8cbb33ffe11756b5e8a928aaf763484c3.